chore: reorganize mutator code
